chore: sync with v1.2.0

+++ lightllm/common/basemodel/prefill_cuda_graph.py
@@ -176,7 +176,7 @@ class PrefillCudaGraph:
         for handle_token_num in self.graph_handle_token_nums[::-1]:
             logger.info(f"Capture prefill cudagraph, handle_token_num: {handle_token_num}")
             total_token_num = handle_token_num
-            input_ids = torch.tensor([1 for _ in range(total_token_num)], dtype=torch.int32, device=self.target_device)
+            input_ids = torch.tensor([1 for _ in range(total_token_num)], dtype=torch.int64, device=self.target_device)
             mem_indexes = model.mem_manager.alloc(len(input_ids)).to(device=self.target_device, non_blocking=True)
             b_req_idx = torch.tensor([model.req_manager.HOLD_REQUEST_ID], dtype=torch.int32, device=self.target_device)
             b_seq_len = torch.empty(1, dtype=torch.int32, device=self.target_device)
@@ -236,7 +236,7 @@ class PrefillCudaGraph:
             for micro_batch_index in [0, 1]:
                 # dummy prefill, capture the cudagraph
                 total_token_num = handle_token_num
-                input_ids = torch.tensor([1 for _ in range(total_token_num)], dtype=torch.int32, device=self.target_device)
+                input_ids = torch.tensor([1 for _ in range(total_token_num)], dtype=torch.int64, device=self.target_device)
                 mem_indexes = model.mem_manager.alloc(len(input_ids)).to(device=self.target_device, non_blocking=True)
                 b_req_idx = torch.tensor([model.req_manager.HOLD_REQUEST_ID], dtype=torch.int32, device=self.target_device)
                 b_seq_len = torch.empty(1, dtype=torch.int32, device=self.target_device)
